+++ servers/env/storage.js
@@ -16,31 +16,36 @@
 	StorageServer.prototype = Object.create(local.env.Server.prototype);
 
 	StorageServer.prototype.handleHttpRequest = function(request, response) {
-		var self = this;
-		var router = local.http.ext.router(request);
-		var respond = local.http.ext.responder(response);
-		router.mp ('HEAD', '/',         __httpListCollections.bind(self, request, respond));
-		router.mpa('GET',  '/', /json/, __httpListCollections.bind(self, request, respond));
-		router.mpt('POST', '/', /json/, __httpGenUniqueCollection.bind(self, request, respond));
-		router.p(__path('^/:collection/?$'), function(match) {
-			var cid = match.path[1];
-			router.m ('HEAD',         __httpGetCollection.bind(self, request, respond, cid));
-			router.ma('GET',  /json/, __httpGetCollection.bind(self, request, respond, cid));
-			router.mt('POST', /json/, __httpAddItem.bind(self, request, respond, cid));
-			router.m ('DELETE',       __httpDeleteCollection.bind(self, request, respond, cid));
-			router.error(response, 'path');
-		});
-		router.p(__path('^/:collection/:item/?$'), function(match) {
-			var cid = match.path[1];
-			var iid = match.path[2];
-			router.m ('HEAD',          __httpGetItem.bind(self, request, respond, cid, iid));
-			router.ma('GET',   /json/, __httpGetItem.bind(self, request, respond, cid, iid));
-			router.mt('PUT',   /json/, __httpSetItem.bind(self, request, respond, cid, iid));
-			router.mt('PATCH', /json/, __httpUpdateItem.bind(self, request, respond, cid, iid));
-			router.m ('DELETE',        __httpDeleteItem.bind(self, request, respond, cid, iid));
-			router.error(response, 'path');
-		});
-		router.error(response);
+		// :DEBUG: temporary helper fn
+		var handled = false, self = this;
+		function route(method, path, fn) {
+			if (handled) return;
+			if (method && path) {
+				path = __path(path);
+				if (path.test(request.path) && RegExp('^'+method+'$','i').test(request.method)) {
+					handled = true;
+					var match = path.exec(request.path);
+					// (request, response, match1, match2, match3...)
+					var args = [request, response].concat(match.slice(1));
+					fn.apply(self, args);
+				}
+			} else
+				response.writeHead(404,'not found').end();
+		}
+
+		route('HEAD',   '^/?$', __httpListCollections);
+		route('GET',    '^/?$', __httpListCollections);
+		route('POST',   '^/?$', __httpGenUniqueCollection);
+		route('HEAD',   '^/:collection/?$', __httpGetCollection);
+		route('GET',    '^/:collection/?$', __httpGetCollection);
+		route('POST',   '^/:collection/?$', __httpAddItem);
+		route('DELETE', '^/:collection/?$', __httpDeleteCollection);
+		route('HEAD',   '^/:collection/:item/?$', __httpGetItem);
+		route('GET',    '^/:collection/:item/?$', __httpGetItem);
+		route('PUT',    '^/:collection/:item/?$', __httpSetItem);
+		route('PATCH',  '^/:collection/:item/?$', __httpUpdateItem);
+		route('DELETE', '^/:collection/:item/?$', __httpDeleteItem);
+		route();
 	};
 
 	// gets (or creates) the collection and returns the keys of its items
@@ -118,119 +123,132 @@
 	}
 
 	function __buildServiceHeaders() {
-		var headerer = local.http.ext.headerer();
-		headerer.addLink('/', 'self current');
-		headerer.addLink('/{title}', 'collection');
+		var headers = {
+			link:[
+				{ href:'/', rel:'self current' },
+				{ href:'/{title}', rel:'collection' }
+			]
+		};
 		Object.keys(this.collections).forEach(function(cid) {
-			headerer.addLink('/'+cid, 'collection', { title:cid });
+			headers.link.push({ href:'/'+cid, rel:'collection', title:cid });
 		});
-		return headerer;
+		return headers;
 	}
 
 	function __buildCollectionHeaders(cid) {
-		var headerer = local.http.ext.headerer();
-		headerer.addLink('/', 'up via service');
+		var headers = {
+			link:[{ href:'/', rel:'up via service' }]
+		};
 		if (cid) {
-			headerer.addLink('/'+cid, 'self current');
-			headerer.addLink('/'+cid+'/{title}', 'item');
+			headers.link.push({ href:'/'+cid, rel:'self current' });
+			headers.link.push({ href:'/'+cid+'/{title}', rel:'item' });
 		}
-		return headerer;
+		return headers;
 	}
 
 	function __buildItemHeaders(cid, iid) {
-		var headerer = local.http.ext.headerer();
-		headerer.addLink('/', 'via service');
+		var headers = {
+			link:[{ href:'/', rel:'via service' }]
+		};
 		if (cid)
-			headerer.addLink('/'+cid, 'up collection');
+			headers.link.push({ href:'/'+cid, rel:'up collection' });
 		if (iid)
-			headerer.addLink('/'+cid+'/'+iid, 'self current');
-		return headerer;
+			headers.link.push({ href:'/'+cid+'/'+iid, rel:'self current' });
+		return headers;
 	}
 
 	// GET /
-	function __httpListCollections(request, respond) {
-		if (/get/i.test(request.method))
-			respond.ok('json', __buildServiceHeaders.call(this)).end(this.collections);
-		else
-			respond.ok(null, __buildServiceHeaders.call(this)).end();
+	function __httpListCollections(request, response) {
+		var headers = __buildServiceHeaders.call(this);
+		if (/get/i.test(request.method)) {
+			headers['content-type'] = 'application/json';
+			response.writeHead(200, 'ok', headers).end(this.collections);
+		} else
+			response.writeHead(200, 'ok', headers).end();
 	}
 
 	// POST /
-	function __httpGenUniqueCollection(request, respond) {
+	function __httpGenUniqueCollection(request, response) {
 		var cid;
 		do { cid = guid(); } while (typeof this.collections[cid] != 'undefined');
 		this.collections[cid] = []; // now defined, not available
 
 		var headers = __buildServiceHeaders.call(this);
 		headers.location = '/'+cid;
-		respond.created('json', headers).end({ id:cid });
+		headers['content-type'] = 'application/json';
+		response.writeHead(201, 'created', headers).end({ id:cid });
 	}
 
 	// GET /:collection
-	function __httpGetCollection(request, respond, cid) {
-		if (/get/i.test(request.method))
-			respond.ok('json', __buildCollectionHeaders.call(this, cid)).end(this.listCollectionItems(cid));
-		else
-			respond.ok(null, __buildCollectionHeaders.call(this, cid)).end();
+	function __httpGetCollection(request, response, cid) {
+		var headers = __buildCollectionHeaders.call(this, cid);
+		if (/get/i.test(request.method)) {
+			headers['content-type'] = 'application/json';
+			response.writeHead(200, 'ok', headers).end(this.listCollectionItems(cid));
+		} else
+			response.writeHead(200, 'ok', headers).end();
 	}
 
 	// POST /:collection
-	function __httpAddItem(request, respond, cid) {
+	function __httpAddItem(request, response, cid) {
 		if (!request.body || typeof request.body != 'object')
-			return respond.unprocessableEntity().end('request body required as a JSON object');
+			return response.writeHead(422, 'unprocessable entity').end('request body required as a JSON object');
 
 		this.setItem(cid, request.body);
 		var headers = __buildCollectionHeaders.call(this, cid);
 		headers.location = '/'+cid+'/'+request.body.id;
-		respond.created('json', headers).end({ id:request.body.id });
+		headers['content-type'] = 'application/json';
+		response.writeHead(201, 'created', headers).end({ id:request.body.id });
 	}
 
 	// DELETE /:collection
-	function __httpDeleteCollection(request, respond, cid) {
+	function __httpDeleteCollection(request, response, cid) {
 		this.removeCollection(cid);
-		respond.noContent(null, __buildCollectionHeaders.call(this)).end();
+		response.writeHead(204, 'no content', __buildCollectionHeaders.call(this, cid)).end();
 	}
 
 	// GET /:collection/:id
-	function __httpGetItem(request, respond, cid, iid) {
+	function __httpGetItem(request, response, cid, iid) {
+		var headers = __buildItemHeaders.call(this, cid, iid);
 		var item = this.getItem(cid, iid);
 		if (item) {
-			if (/get/i.test(request.method))
-				respond.ok('json', __buildItemHeaders.call(this, cid, iid)).end(item);
-			else
-				respond.ok(null, __buildItemHeaders.call(this, cid, iid)).end();
+			if (/get/i.test(request.method)) {
+				headers['content-type'] = 'application/json';
+				response.writeHead(200, 'ok', headers).end(item);
+			} else
+				response.writeHead(200, 'ok', headers).end();
 		} else
-			respond.notFound(null, __buildItemHeaders.call(this, cid)).end();
+			response.writeHead(404, 'not found', headers).end();
 	}
 
 	// PUT /:collection/:id
-	function __httpSetItem(request, respond, cid, iid) {
+	function __httpSetItem(request, response, cid, iid) {
 		if (!request.body || typeof request.body != 'object')
-			return respond.unprocessableEntity().end('request body required as a JSON object');
+			return response.writeHead(422, 'unprocessable entity').end('request body required as a JSON object');
 
 		request.body.id = iid;
 		this.setItem(cid, request.body);
-		respond.noContent(null, __buildItemHeaders.call(this, cid, iid)).end();
+		response.writeHead(204, 'no content', __buildItemHeaders.call(this, cid, iid)).end();
 	}
 
 	// PATCH /:collection/:id
-	function __httpUpdateItem(request, respond, cid, iid) {
+	function __httpUpdateItem(request, response, cid, iid) {
 		if (!request.body || typeof request.body != 'object')
-			return respond.unprocessableEntity().end('request body required as a JSON object');
+			return response.writeHead(422, 'unprocessable entity').end('request body required as a JSON object');
 
 		var item = this.getItem(cid, iid);
 		if (item) {
 			item = patch(item, request.body);
 			this.setItem(cid, item);
-			respond.noContent(null, __buildItemHeaders.call(this, cid, iid)).end();
+			response.writeHead(204, 'no content', __buildItemHeaders.call(this, cid, iid)).end();
 		} else
-			respond.notFound(null, __buildItemHeaders.call(this, cid)).end();
+			response.writeHead(404, 'not found', __buildItemHeaders.call(this, cid, iid)).end();
 	}
 
 	// DELETE /:collection/:id
-	function __httpDeleteItem(request, respond, cid, iid) {
+	function __httpDeleteItem(request, response, cid, iid) {
 		this.removeItem(cid, iid);
-		respond.noContent(null, __buildItemHeaders.call(this, cid)).end();
+		response.writeHead(204, 'no content', __buildItemHeaders.call(this, cid, iid)).end();
 	}
 
 	// brings updates into org value